requirements.txt
asgiref==3.3.4
autopep8==1.5.7
beautifulsoup4==4.9.1
dj-database-url==0.5.0
Django==3.2.3
django-bootstrap4==2.2.0
django-crispy-forms==1.9.2
django-js-asset==1.2.2
django-mptt==0.12.0
importlib-metadata==1.7.0
Pillow==7.2.0
psycopg2==2.8.5
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
soupsieve==2.0.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
zipp==3.1.0
